requirements.txt
amulet-core==1.9.25
amulet-leveldb==1.0.2
amulet-nbt==2.1.3
lz4==4.3.3
mutf8==1.0.6
numpy==1.26.4
platformdirs==3.11.0
portalocker==2.10.0
PyMCTranslate==1.2.27

zstd~=1.5.5.1